+++ jetpack_vendor/i18n-map.php
@@ -46,11 +46,11 @@ return array(
     ),
     'jetpack-masterbar' => array(
       'path' => 'jetpack_vendor/automattic/jetpack-masterbar',
-      'ver' => '0.14.0-alpha1742222219',
+      'ver' => '0.14.0-alpha1742222296',
     ),
     'jetpack-mu-wpcom' => array(
       'path' => 'jetpack_vendor/automattic/jetpack-mu-wpcom',
-      'ver' => '6.2.0-alpha1742222219',
+      'ver' => '6.2.0-alpha1742222296',
     ),
     'jetpack-password-checker' => array(
       'path' => 'jetpack_vendor/automattic/jetpack-password-checker',
@@ -67,6 +67,10 @@ return array(
     'jetpack-stats-admin' => array(
       'path' => 'jetpack_vendor/automattic/jetpack-stats-admin',
       'ver' => '0.24.4',
+    ),
+    'jetpack-subscribers-dashboard' => array(
+      'path' => 'jetpack_vendor/automattic/jetpack-subscribers-dashboard',
+      'ver' => '0.1.0-alpha1742222296',
     ),
     'jetpack-sync' => array(
       'path' => 'jetpack_vendor/automattic/jetpack-sync',